Proposed split-up of quickflat module. Bish.
